+++ src/main/java/cart/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 import java.util.HashMap;
 import java.util.Map;
