+++ src/main/java/cart/dto/response/Response.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 import lombok.AccessLevel;
 import lombok.Getter;
