+++ src/main/java/cart/dto/response/ResultResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 import lombok.Getter;
